backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-ai
langchain
langchain-google-genai
httpx
python-jose[cryptography]
python-dotenv